requirements-dev.txt
Django==2.0.9
dj-database-url==0.5.0
dj-static==0.0.6
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
mysqlclient==1.4.2.post1
Pillow==6.0.0
python-decouple==3.1
pytz==2019.1
static3==0.7.0